requirements.txt
# ─────────────────────────────────────────────────
# 약사 챗봇 RAG - Streamlit 앱 의존성 목록
# 설치: pip install -r requirements.txt
# ─────────────────────────────────────────────────

# Streamlit UI
streamlit>=1.35.0

# FastAPI Backend
fastapi>=0.111.0
uvicorn>=0.30.0

# LangChain 핵심
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0

# Vector DB
chromadb>=0.5.0

# 임베딩 모델 (BGE-M3)
sentence-transformers>=3.0.0

# BM25 검색
rank-bm25>=0.2.2

# 한국어 형태소 분석
kiwipiepy>=0.18.0

# LLM API
openai>=1.30.0

# RAGAS 평가
ragas>=0.1.14
datasets>=2.19.0

# 유틸리티
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0

# GPU (RTX 2070 – CUDA 12.1 기준)
# 아래 커맨드로 CUDA 버전 PyTorch를 먼저 설치하세요:
#   pip install torch>=2.3.0 --index-url https://download.pytorch.org/whl/cu121
# 그 다음 나머지 패키지 설치:
#   pip install -r requirements.txt
torch>=2.3.0
